Update to work with the latest version of 'richtext'
The `richtext` ingredient now uses TinyMCE 7.9.1
(https://github.com/eprints/richtext/pull/10) so this changes
`initTinyMCE_for_easy_pages` to match the changes in the version bump.

The `textpattern` ingredient has been removed and I believe is just
included by default now so I have just removed that.

The `preview_with_css_branding` is a modified version of `preview` so I
have pulled the new version of `preview` (as the previous version was
for 4.6.1) and then re-applied the modifications (filter content with
`preview_substitutions` and add a font style-sheet).