Automatically set the width and prevent it from being altered
If we don't set the width it will automatically fill the box its in
which seems to make logical sense and it feels nicer to then not be able
to change the width as otherwise you accidentally shift it one way or
the other while trying to adjust the height.

I originally noticed this because some of the buttons (link, image, code
and preview) were being hidden in a dropdown menu by default which
didn't make much sense.